correcion test clase juego

+++ back/src/test/java/es/santander/ascender/proyecto20/JuegoTest.java
@@ -36,7 +36,7 @@ public class JuegoTest {
 
     @Test
     void testInicializarJuego() {
-        int sesion = jueguito.inicializarJuego();
+        long sesion = jueguito.inicializarJuego();
         assertEquals(1, sesion);
     }
